+++ src/app/(non-navbar)/email-signin/_component/SignupErrorText.tsx
