+++ src/app/(non-navbar)/email-signin/_component/layout.tsx
@@ -1,0 +1,19 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Let's",
+  icons: {
+    icon: "./assets/mainLogo.svg",
+  },
+};
+
+const EmailSigninLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <main>
+      {children}
+      <div id="portal" />
+    </main>
+  );
+};
+
+export default EmailSigninLayout;
