+++ src/app/(non-navbar)/email-signin/layout.tsx
