+++ src/app/(non-navbar)/email-signup/_component/SignupEmailInfo.tsx
