+++ src/app/(non-navbar)/email-signup/_component/SignupStepTwo.tsx
@@ -1,5 +1,5 @@
 import Button from "@/app/_component/common/atom/Button";
-import SigninInput from "@/app/email-signin/_component/SigninInput";
+import SigninInput from "@/app/(non-navbar)/email-signin/_component/SigninInput";
 
 interface Props {
   setStep: React.Dispatch<React.SetStateAction<number>>;
@@ -11,7 +11,7 @@ const SignupStepTwo = ({ setStep }: Props) => {
   };
 
   return (
-    <div className="flex flex-col items-center mt-11 h-full px-6">
+    <div className="flex flex-col items-center pt-11 h-full px-6">
       <SigninInput
         id="email-auth"
         name="email-auth"
